Revert "Revert "Rename, recombine TFRecordInput class.""
This reverts commit 0034724a080ddfd1dec39424fc44c9529d3dbb2e.
