+++ experiments/tf_trainer/common/tfrecord_input_test.py
@@ -18,17 +18,78 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import numpy as np
+import tensorflow as tf
+
 from tf_trainer.common import base_model
 from tf_trainer.common import tfrecord_input
 from tf_trainer.common import types
 
-import numpy as np
-import tensorflow as tf
-
 
 FLAGS = tf.app.flags.FLAGS
 
+
 class TFRecordInputTest(tf.test.TestCase):
+
+  def setUp(self):
+    FLAGS.text_feature = 'comment'
+    ex = tf.train.Example(
+      features=tf.train.Features(
+          feature={
+              'label':
+                  tf.train.Feature(
+                      float_list=tf.train.FloatList(value=[0.8])),
+              'ignored-label':
+                  tf.train.Feature(
+                      float_list=tf.train.FloatList(value=[0.125])),
+              'comment':
+                  tf.train.Feature(
+                      bytes_list=tf.train.BytesList(
+                          value=['Hi there Bob'.encode('utf-8')]))
+                  }))
+    self.ex_tensor = tf.convert_to_tensor(ex.SerializeToString(),
+                                          dtype=tf.string)
+
+  def test_TFRecordInput_unrounded(self):
+    FLAGS.round_labels = False
+    FLAGS.labels = 'label'
+    dataset_input = tfrecord_input.TFRecordInput()
+
+    with self.test_session():
+      features, labels = dataset_input._read_tf_example(self.ex_tensor)
+      self.assertEqual(features[base_model.TEXT_FEATURE_KEY].eval(),
+                       b'Hi there Bob')
+      np.testing.assert_almost_equal(labels['label'].eval(), 0.8)
+      self.assertEqual(list(labels), ['label'])
+
+  def test_TFRecordInput_default_values(self):
+    FLAGS.labels = 'label,fake_label'
+    FLAGS.round_labels = False
+    dataset_input = tfrecord_input.TFRecordInput()
+
+    with self.test_session():
+      features, labels = dataset_input._read_tf_example(self.ex_tensor)
+      self.assertEqual(features[base_model.TEXT_FEATURE_KEY].eval(),
+                       b'Hi there Bob')
+      np.testing.assert_almost_equal(labels['label'].eval(), 0.8)
+      np.testing.assert_almost_equal(labels['fake_label'].eval(), -1.0)
+
+  def test_TFRecordInput_rounded(self):
+    FLAGS.labels = 'label'
+    FLAGS.round_labels = True
+    dataset_input = tfrecord_input.TFRecordInput()
+
+    with self.test_session():
+      features, labels = dataset_input._read_tf_example(self.ex_tensor)
+      self.assertEqual(features[base_model.TEXT_FEATURE_KEY].eval(),
+                       b'Hi there Bob')
+      np.testing.assert_almost_equal(labels['label'].eval(), 1.0)
+
+if __name__ == '__main__':
+  tf.test.main()
+
+
+class TFRecordInputWithTokenizerTest(tf.test.TestCase):
 
   def setUp(self):
     FLAGS.text_feature = "comment"
@@ -66,10 +127,10 @@ class TFRecordInputTest(tf.test.TestCase):
                        [12, 13, 999])
       self.assertAlmostEqual(labels["label"].eval(), 0.8)
 
-  def test_TFRecordInput_default_values(self):
+  def test_TFRecordInputWithTokenizer_default_values(self):
     FLAGS.labels = "label,fake_label"
     FLAGS.round_labels = False
-    dataset_input = tfrecord_input.TFRecordInput(
+    dataset_input = tfrecord_input.TFRecordInputWithTokenizer(
         train_preprocess_fn=self.preprocessor)
 
     with self.test_session():
@@ -79,10 +140,10 @@ class TFRecordInputTest(tf.test.TestCase):
       self.assertAlmostEqual(labels["label"].eval(), 0.8)
       self.assertAlmostEqual(labels["fake_label"].eval(), -1.0)
 
-  def test_TFRecordInput_rounded(self):
+  def test_TFRecordInputWithTokenizer_rounded(self):
     FLAGS.labels = "label"
     FLAGS.round_labels = True
-    dataset_input = tfrecord_input.TFRecordInput(
+    dataset_input = tfrecord_input.TFRecordInputWithTokenizer(
         train_preprocess_fn=self.preprocessor)
 
     with self.test_session():
